@@ -18,9 +18,10 @@
     "@babel/core": "^7.14.2",
     "@babel/preset-env": "^7.14.2",
     "@babel/preset-react": "^7.13.13",
-    "@testing-library/jest-dom": "^5.12.0",
-    "@testing-library/react": "^11.2.7",
-    "@types/jest": "^26.0.23",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/jest": "^26.0.24",
     "babel-jest": "^26.6.3",
     "babel-loader": "^8.2.2",
     "codeceptjs": "^3.0.7",
@@ -39,10 +40,10 @@
     "webpack-dev-server": "^4.8.1"
   },
   "dependencies": {
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.4",
-    "redux": "^4.1.0",
+    "react": "^18.0.0-rc.3",
+    "react-dom": "^18.0.0-rc.3",
+    "react-redux": "^7.2.8",
+    "redux": "^4.2.0",
     "redux-thunk": "^2.3.0"
   }
 }
