@@ -8,7 +8,9 @@ module.exports = {
         },
       },
     ],
-    '@babel/preset-react',
+    ['@babel/preset-react', {
+      runtime: 'automatic',
+    }],
   ],
   plugins: [
     [
