@@ -1,5 +1,14 @@
 export default function App() {
   return (
-    <></>
+    <div>
+      Restaurant
+      <div>
+        <ul>
+          <li>서울</li>
+          <li>부산</li>
+          <li>독도</li>
+        </ul>
+      </div>
+    </div>
   );
 }
